Make TextDropdown generic
Da kan den brukes flere steder.
Den har dukket opp i flere skisser nå, så da blir den nok snart tatt i bruk.
